requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
distro==1.9.0
faiss-cpu==1.8.0.post1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
numpy==1.26.4
openai==1.41.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
setuptools==75.1.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
